refactor: organize lightcurve API routes by response format
