+++ multisurveys-apis/src/lightcurve_api/api.py
@@ -3,7 +3,8 @@ from fastapi.middleware.cors import CORSMiddleware
 from fastapi.middleware.gzip import GZipMiddleware
 from fastapi.staticfiles import StaticFiles
 from prometheus_fastapi_instrumentator import Instrumentator
-from .routes import rest, conesearch
+from .routes.htmx import lightcurve as htmx_lightcurve
+from .routes.json import conesearch, lightcurve as json_lightcurve
 
 app = FastAPI(openapi_url="/v2/lightcurve/openapi.json")
 instrumentator = Instrumentator().instrument(app).expose(app)
@@ -19,7 +20,7 @@ app.add_middleware(
 
 app.add_middleware(GZipMiddleware, minimum_size=1000)
 
-app.include_router(rest.router)
+app.include_router(json_lightcurve.router)
 app.include_router(conesearch.router)
 
 app.mount(
