+++ multisurveys-apis/src/lightcurve_api/routes/json/lightcurve.py
@@ -5,7 +5,7 @@ from fastapi import APIRouter, HTTPException, Query
 
 from core.config.dependencies import db_dependency
 
-from ..services.lightcurve_service import (
+from ...services.lightcurve_service import (
     get_detections,
     get_detections_by_list,
     get_forced_photometry,
@@ -13,7 +13,7 @@ from ..services.lightcurve_service import (
     get_non_detections,
     get_non_detections_by_list,
 )
-from ..services.validations import survey_validate
+from ...services.validations import survey_validate
 
 router = APIRouter()
